requirements.txt
aiofiles==24.1.0
aiosqlite==0.20.0
asyncua==1.1.5
cffi==1.17.1
cryptography==43.0.1
opcua-widgets==0.6.1
pycparser==2.22
pyOpenSSL==24.2.1
PyQt5==5.15.11
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sortedcontainers==2.4.0
typing_extensions==4.12.2
numpy==1.24.4
pyqtgraph==0.13.3
qasync==0.27.1
